# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My User Project CRUD",
    "description": "My User Project Application API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:8000",
  "basePath": "/",
  "tags": [
    {
      "name": "Users",
      "description": "API for users in the system"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/users": {
      "get": {
        "summary": "Get all the users",
        "description": "Get all the users",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array"
            }
          }
        }
      }
    },
    "/users/{getId}": {
      "get": {
        "summary": "Get user by id",
        "description": "Get user by id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "getId",
            "in": "path",
            "description": "get user by id ",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
    "/users/{deleteId}": {
      "delete": {
        "summary": "delete one user",
        "description": "delete one user by id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "deleteId",
            "in": "path",
            "description": "user by id that needs to be deleted",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
    "/users/{updateId}": {
      "put": {
        "summary": "Update user by id",
        "description": "Update user by id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "updateId",
            "in": "path",
            "description": "update user by id",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "user object",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "user": {
                  "type": "object"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    }
  }
}